Add watermark, serverside generation, upload to s3
This patch adds a watermark to the homepage so that users are
incentivized to click through and generate a linkable image. We also
added a lambda to actually generate the image serverside and upload to
s3. Instead of using a database and an auto-incrementing number, like in
previous versions of BCS, we just generate a random string so that we
don't have to worry about having a database.

In the future we should probably verify that the random string doesn't
collide with an existing image. But that only happens 1 in 26^10 times.
So I think we should be good.