Add OpenGraph, utilize stale-while-revalidate
This patch modifies the swag page to use a serverless function instead
of static file that updates based on the URL parameter. This allows us
to properly support opengraph headers. To ensure that we don't incur too
much cost and can remain fast, we add the stale-while-revalidate cache
control header with Zeit to ensure frequently hit pages aren't invoking
the servless function all that much.

This also gets rid of the deprecated warning in next.js by properly
using the serverless function.